Implementation a new MVI pattern
Jake did a nice talk on managing state with RxJava.
I liked the idea and tried implementing it. Still
working on it.
Video is: https://youtu.be/0IKHxjkgop4

Activity emmits Intents which are translated into
Actions, their Results are then translated back
to States and we render them inside the Activity.

I am still missing management of rotation change.