Build notification when service is kicked
Replaced the receiver by the service for we are
running stuff in the background so seemed more
appropriate.
Using the basic logo so the design is nothing
great.
The notification can still happens for old or
currently taking place matches so still work to
do. Also for notification deletion, disable FAB,
etc.