Playing with Sticky headers
Code inspired from https://github.com/edubarr/header-decor

The code is ugly now as the logic forces you to manage two list,
items + headers and you would need to find easily each item's
header based on position.
I'd rather have only one list with only a way to know if a item is
a header or not. Not sure how if it's feasable using item decoration but
maybe worth the try.